// src/app/models/store.ts
export interface StoreSettings {
    // address:string
    // email:string
    // number:string
    // fax:string
    // description:string
    // facebook_url:string
    // instagram_url:string
    // twitter_url:string
    // stock_level:number
    min_sub: number
    unlimited: number 
    delivery: number 
    bidvalue: number
    commission: number
}